chore: move the diff function
